Remove warnings
