+++ gradle/libs.versions.toml
@@ -58,8 +58,8 @@ playServicesAuth = "21.3.0"
 playServicesOssLicenses = "17.1.0"
 # Stay on 4.26.1 due to https://github.com/firebase/firebase-android-sdk/issues/5997
 protobuf = "4.26.1"
-protobuf-gen-grpc-java = "1.72.0"
-protobuf-gen-grpc-kotlin = "1.4.1:jdk8@jar"
+protobuf-gen-grpc-java = "1.71.0"
+protobuf-gen-grpc-kotlin = "1.3.0:jdk8@jar"
 protobuf-gen-javalite = "3.0.0"
 roborazzi = "1.43.1"
 room = "2.7.1"
