+++ media/data/src/main/java/com/google/android/horologist/media/data/datasource/MediaDownloadLocalDataSource.kt
@@ -17,12 +17,12 @@
 package com.google.android.horologist.media.data.datasource
 
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
-import com.google.android.horologist.media.data.database.dao.MediaDownloadDao
-import com.google.android.horologist.media.data.database.dao.MediaDownloadDao.Companion.DOWNLOAD_PROGRESS_END
-import com.google.android.horologist.media.data.database.dao.MediaDownloadDao.Companion.DOWNLOAD_PROGRESS_START
-import com.google.android.horologist.media.data.database.dao.MediaDownloadDao.Companion.SIZE_UNKNOWN
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntity
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntityStatus
+import com.google.android.horologist.media.database.dao.MediaDownloadDao
+import com.google.android.horologist.media.database.dao.MediaDownloadDao.Companion.DOWNLOAD_PROGRESS_END
+import com.google.android.horologist.media.database.dao.MediaDownloadDao.Companion.DOWNLOAD_PROGRESS_START
+import com.google.android.horologist.media.database.dao.MediaDownloadDao.Companion.SIZE_UNKNOWN
+import com.google.android.horologist.media.database.model.MediaDownloadEntity
+import com.google.android.horologist.media.database.model.MediaDownloadEntityStatus
 import com.google.android.horologist.media.model.MediaDownload
 import kotlinx.coroutines.flow.Flow
