+++ media/data/src/main/java/com/google/android/horologist/media/data/mapper/MediaDownloadStatusMapper.kt
@@ -17,8 +17,8 @@
 package com.google.android.horologist.media.data.mapper
 
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntity
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntityStatus
+import com.google.android.horologist.media.database.model.MediaDownloadEntity
+import com.google.android.horologist.media.database.model.MediaDownloadEntityStatus
 import com.google.android.horologist.media.model.MediaDownload
 
 /**
