+++ media/data/src/main/java/com/google/android/horologist/media/data/service/download/DownloadManagerListener.kt
@@ -21,8 +21,8 @@ import androidx.media3.exoplayer.offline.Download
 import androidx.media3.exoplayer.offline.DownloadManager
 import androidx.media3.exoplayer.offline.DownloadService
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
-import com.google.android.horologist.media.data.database.mapper.MediaDownloadEntityStatusMapper
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntityStatus
+import com.google.android.horologist.media.database.mapper.MediaDownloadEntityStatusMapper
+import com.google.android.horologist.media.database.model.MediaDownloadEntityStatus
 import com.google.android.horologist.media.data.datasource.MediaDownloadLocalDataSource
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
