+++ media/data/src/main/java/com/google/android/horologist/media/database/model/PlaylistEntity.kt
@@ -14,23 +14,19 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.media.data.database.model
+package com.google.android.horologist.media.database.model
 
 import androidx.room.Entity
-import androidx.room.Index
+import androidx.room.PrimaryKey
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
 
 /**
- * Cross-reference table that holds relationship between [PlaylistEntity] and [MediaEntity].
+ * A table to store playlist information.
  */
 @ExperimentalHorologistApi
-@Entity(
-    primaryKeys = ["playlistId", "mediaId"],
-    indices = [
-        Index(value = ["mediaId"]),
-    ],
-)
-public data class PlaylistMediaEntity(
-    val playlistId: String,
-    val mediaId: String,
+@Entity
+public data class PlaylistEntity(
+    @PrimaryKey val playlistId: String,
+    val name: String,
+    val artworkUri: String?,
 )
