+++ media/data/src/test/java/com/google/android/horologist/media/data/database/mapper/MediaDownloadEntityStatusMapperTest.kt
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.media.data.database.mapper
+package com.google.android.horologist.media.database.mapper
 
 import androidx.media3.exoplayer.offline.Download
-import com.google.android.horologist.media.data.database.model.MediaDownloadEntityStatus
+import com.google.android.horologist.media.database.model.MediaDownloadEntityStatus
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
