+++ media/data/src/test/java/com/google/android/horologist/media/data/mapper/PlaylistMapperTest.kt
@@ -16,9 +16,9 @@
 
 package com.google.android.horologist.media.data.mapper
 
-import com.google.android.horologist.media.data.database.model.MediaEntity
-import com.google.android.horologist.media.data.database.model.PlaylistEntity
-import com.google.android.horologist.media.data.database.model.PopulatedPlaylist
+import com.google.android.horologist.media.database.model.MediaEntity
+import com.google.android.horologist.media.database.model.PlaylistEntity
+import com.google.android.horologist.media.database.model.PopulatedPlaylist
 import com.google.android.horologist.media.model.Media
 import com.google.android.horologist.media.model.Playlist
 import com.google.common.truth.Truth.assertThat
